--- class-components/src/interfaces/errorBounderyTypes/errorBounderyTypes.ts
@@ -1,11 +1,0 @@
-import { ReactNode } from 'react'
-
-interface BounderyProps {
-  children: ReactNode
-}
-interface BounderyState {
-  hasError: boolean
-  error: Error | null
-}
-
-export type { BounderyProps, BounderyState }
