--- class-components/src/utils/renderTests.tsx
@@ -1,8 +1,0 @@
-import { MemoryRouter } from 'react-router-dom'
-import { ReactElement, ReactNode } from 'react'
-
-const renderTestsComponent = <P,>(component: ReactElement<P>): ReactNode => (
-  <MemoryRouter>{component}</MemoryRouter>
-)
-
-export default renderTestsComponent
